+++ ckanext/datasetsnippets/plugin.py
@@ -29,6 +29,8 @@ class DatasetsnippetsPlugin(plugins.SingletonPlugin):
         # we introduce a new config setting here to be able to have
         # different behaviour in snippets and regular CKAN UI
         config['datasetsnippets.datasets_per_page'] = config.get('datasetsnippets.datasets_per_page', 10)
+        config['datasetsnippets.default_root_breadcrumb'] = config.get(
+        'datasetsnippets.default_root_breadcrumb', 'Berlin Open Data')
 
         path = os.path.abspath(__file__)
         dir_path = os.path.dirname(path)